+++ apps/tests/conftest.py
@@ -1,6 +1,8 @@
 import pytest
+from django.core.files.uploadedfile import SimpleUploadedFile
 from rest_framework.test import APIClient
 
+from apps.tests.application_json_mock import JSON_data
 from apps.tests.data import create_user
 
 
@@ -15,6 +17,11 @@ def api_client_user() -> APIClient:
     client = APIClient()
     client.force_login(user=user)
 
-    yield client
+    yield client, user
 
     client.logout()
+
+
+@pytest.fixture
+def mocked_application_file() -> SimpleUploadedFile:
+    return SimpleUploadedFile('application_dump.json', str.encode(JSON_data))